backend/requirements.txt
Flask==3.0.2
Flask-Cors==4.0.0
python-dotenv==1.0.1
requests==2.31.0
pytest==8.0.2
black==24.2.0
flake8==7.0.0
python-jose==3.3.0
gunicorn==21.2.0 